+++ apirest-springboot/src/main/java/com/prelev/apirest_springboot/dto/PrelevementJourDTO.java
@@ -1,12 +1,14 @@
 package com.prelev.apirest_springboot.dto;
 
+import java.math.BigDecimal;
+
 public class PrelevementJourDTO {
     private String nom;
-    private Integer prix;
+    private BigDecimal prix;
     private int jour;
 
     // Constructeur
-    public PrelevementJourDTO(String nom, int prix, int jour) {
+    public PrelevementJourDTO(String nom, BigDecimal prix, int jour) {
         this.nom = nom;
         this.prix = prix;
         this.jour = jour;
@@ -16,8 +18,8 @@ public class PrelevementJourDTO {
     public String getNom() { return nom; }
     public void setNom(String nom) { this.nom = nom; }
 
-    public Integer getPrix() { return prix; }
-    public void setPrix(Integer prix) { this.prix = prix; }
+    public BigDecimal getPrix() { return prix; }
+    public void setPrix(BigDecimal prix) { this.prix = prix; }
 
     public int getJour() { return jour; }
     public void setJour(int jour) { this.jour = jour; }
